New process.exceptionCatcher is inherited by callbacks.
This allows a web framework to catch at the top level exceptions
thrown by page dispatch code, and to return a 500 Internal Server Error
for that request without disturbing other requests. This works
even if the page dispatch code is async.

FatalException tries process.exceptionCatcher first before
falling back to emitting an "uncaughtException" event.

Whenever a callback is registered, EventEmitter.addListener()
sets callback._exceptionCatcher to the current value of
process.exceptionCatcher. When the callback is dispatched,
EventEmitter.emit() sets process.exceptionCatcher to the
callback's ._exceptionCatcher.

In this way callbacks registered by callbacks also inherit the
exceptionHandler. It is also possible to chain exceptionHandlers.