Fix compileJS edge case.
Take care of edge cases where alternative module types might provide a loader, but no compiler, and then when compiled as JavaScript, the module exports a string.  Since it's already been parsed as JavaScript, it should not be parsed as JavaScript again, even though it's not already a .js module.