A big base64 file encoding/decoding test
Originally written to try to track down this bug report: http://permalink.gmane.org/gmane.comp.lang.javascript.nodejs/13924

As it turns out, it's showing a completely different kind of failure.  When the file is base64 encoded with the native program, then decoded with node, it doesn't match the original.  Something's fishy.