module: search symlink path node_modules as well
When modules are loaded from a symlink location, prior to this patch,
their resulting `node_modules` lookup path was only based on their
realpath location, not where they were loaded from.

That behavior presents a problem in 2 cases, both of which are addressed
by this patch.

The first problem is the case of peer dependencies that are symlinked
into an application workspace.  Consider a developer who is using a
framework, and is also developing two extensions for that framework.

    app
    └── node_modules
        ├── framework
        ├── addon -> ../../../dev/addon
        └── plugin -> ../../../dev/plugin

When their plugin and addon want to load the framework as a peer
dependency, they are unable to do so, because their
`require('framework')` call will not find it in the app location.

The second problem is that insisting that dependencies always be found
in the modules' realpath-resolved node_modules hierarchy limits the
ability of a module manager to symbolicly link all deps into a shared
global location (for efficiency of disk space, networking, and module
build/install times).

The limitation is that, in such a symlink-in-a-store module management
system, if a developer is working on two applications, then updating the
meta-dependencies in one necessarily update those dependencies in the
other, since the deps must be updated in the centralized module store.

Additionally, any such module store will result in infinite symbolic
link loops within the store in the case of transitive dependency cycles,
as dependencies symlink to one another in their realpath'ed node_modules
folders.  That is, if `A -> B -> A`, then `.store/a/node_modules/b` will
be a symlink to `.store/b`, and `.store/b/node_modules/a` will be a
symlink to `.store/a`, resulting in
`.store/a/node_modules/b/node_modules/a/...`, which trips up some
editors and other development tools.

With this change, there are at least 2 methods whereby a module manager
could implement a central dependency store without requiring "action at
a distance", and without any symbolic link loops.

Consider a module store structure like this:

    .store
    ├── bar
    │   ├── 2.3.4
    │   │   └── index.js
    │   ├── 9.0.0
    │   │   └── index.js
    │   └── 9.8.7
    │       └── index.js
    ├── baz
    │   └── 1.2.5
    │       └── index.js
    └── foo
        └── 1.2.3
            └── index.js

Further consider a dependency structure like so:

    app
    ├── baz@1.2.x
    │   └── bar@2.x
    └── foo@1.2.x
        └── bar@9.x

In the first module manager strategy, we symlink the _contents_ of each
dep into its appropriate folder inside the application, resulting in a
structure like this:

    app
    ├── index.js
    └── node_modules
        ├── baz
        │   ├── index.js -> ../../../.store/baz/1.2.5/index.js
        │   └── node_modules
        │       └── bar
        │           └── index.js -> ../../../../../.store/bar/2.3.4/index.js
        └── foo
            ├── index.js -> ../../../.store/foo/1.2.3/index.js
            └── node_modules
                └── bar
                    └── index.js -> ../../../../../.store/bar/9.0.0/index.js

In the second module manager strategy, each module in the application's
workspace is linked to a `contents` folder, and a `package.json` file
is generated (by the module manager) with the JSON `{"main":"content"}`,
resulting in the following layout on disk:

    app
    ├── index.js
    └── node_modules
        ├── baz
        │   ├── contents -> ../../../.store/baz/1.2.5
        │   ├── node_modules
        │   │   └── bar
        │   │       ├── contents -> ../../../../../.store/bar/2.3.4
        │   │       └── package.json
        │   └── package.json
        └── foo
            ├── contents -> ../../../.store/foo/1.2.3
            ├── node_modules
            │   └── bar
            │       ├── contents -> ../../../../../.store/bar/9.8.7
            │       └── package.json
            └── package.json

With this patch:

A) there is no possibility of symbolic link loops (since symlinks are
not created underneath any symbolicly linked folders)

B) it is possible to have separate copies of `foo` that get their own
different versions of `bar` than other linked copies of `foo` in other
applications.

C) either strategy is possible with the same module store.

D) the peerDependencies symlink problem is resolved, without the need
for a --resolve-symlinks switch.  This simplifies the codebase by
removing several `isMain` checks and a user-facing switch.

E) all current behavior is preserved, so it is safe to add without
putting behind a flag.  All current strategies for module loading work
fine, the singleton nature of modules based on their realpath is not
altered.  (Further investigation should be done to ensure that npm, ied,
pnpm, and the like are unaffected.)

F) no changes are made to any public interfaces; the only thing that
changes is that the `module._paths` array gets extended by a few
entries.

One potential limitation with this approach is that loading a single
module from _multiple different_ symlink locations will still result in
only loading the module once (because the realpath is used as the cache
key).  Thus, if an application links a module into multiple different
locations, and expects that module to end up loading different versions
of its dependencies as a result, then they will find that only the first
dependency resolution is used.

However, this limitation is not a departure from any other module
symlink handling strategies thus proposed.  Removing this limitation
would be a significant change to the de facto contract of the module
loader thusfar, that modules are singletons based on their realpath, and
would therefor be risky to introduce anyway.